identifiers: swap to 64bit identifier
Swap to a 64bit identifier to eliminate any chance that we will run out
of addresses and give us more freedom in assigning sequential IDs.

+++ include/infuse/identifiers.h
@@ -25,9 +25,9 @@ extern "C" {
 /**
  * @brief Get local device ID
  *
- * @return uint32_t local device ID
+ * @return uint64_t local device ID
  */
-uint32_t infuse_device_id(void);
+uint64_t infuse_device_id(void);
 
 /**
  * @}
